OLCTriangle: use a mixed breadth-first/depth-first search for large trees
This won't affect fast problems but can reduce the memory usage for hard
ones.

Experiments with worst-case flights with many gps fixes revealed a
saving of over 2/3 compared to the old strategy (450mb instead of 1.6gb
for a flight with 4096 points). Flights with less points don't have that
problem, due to a nonlinear dependency to flight size.